requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
werkzeug
python-dotenv
pyodbc
